WIP refactor plotting module
